Updated the log message to include the tracer bean being wrapped, and added 'spring' to autoconfigure artifact id and package
