--- opentracing-api-extensions-tracer-autoconfigure/src/main/resources/META-INF/spring.factories
@@ -1,2 +1,0 @@
-org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
-io.opentracing.contrib.api.tracer.autoconfigure.TracerBeanPostProcessor
