+++ opentracing-api-extensions-tracer-spring-autoconfigure/src/main/resources/META-INF/spring.factories
@@ -1,0 +1,2 @@
+org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
+io.opentracing.contrib.api.tracer.spring.autoconfigure.TracerBeanPostProcessor
