+++ opentracing-api-extensions-tracer-spring-autoconfigure/src/test/java/io/opentracing/contrib/api/tracer/spring/autoconfigure/TracerBeanPostProcessorTest.java
@@ -11,41 +11,66 @@
  * or implied. See the License for the specific language governing permissions and limitations under
  * the License.
  */
-package io.opentracing.contrib.api.tracer.autoconfigure;
+package io.opentracing.contrib.api.tracer.spring.autoconfigure;
 
-import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertNotEquals;
 
+import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
+import org.mockito.Matchers;
+import org.mockito.Mockito;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
+
 import io.opentracing.Tracer;
+import io.opentracing.contrib.api.SpanData;
+import io.opentracing.contrib.api.TracerObserver;
 import io.opentracing.mock.MockTracer;
 import io.opentracing.util.ThreadLocalActiveSpanSource;
 
 @SpringBootTest(
-        classes = {TracerBeanPostProcessorNoObserversTest.SpringConfiguration.class})
+        classes = {TracerBeanPostProcessorTest.SpringConfiguration.class})
 @RunWith(SpringJUnit4ClassRunner.class)
-public class TracerBeanPostProcessorNoObserversTest {
+public class TracerBeanPostProcessorTest {
+
+    private static final MockTracer mockTracer = new MockTracer(new ThreadLocalActiveSpanSource());
+
+    private static final TracerObserver tracerObserver = Mockito.mock(TracerObserver.class);
 
     @Configuration
     @EnableAutoConfiguration
     public static class SpringConfiguration {
         @Bean
-        public MockTracer tracer() {
-            return new MockTracer(new ThreadLocalActiveSpanSource());
+        public Tracer tracer() {
+            return mockTracer;
+        }
+
+        @Bean
+        public TracerObserver observer() {
+            return tracerObserver;
         }
     }
 
     @Autowired
     protected Tracer tracer;
 
-    @Test
-    public void testTracerNotWrapped() {
-        assertEquals(MockTracer.class, tracer.getClass());
+    @Before
+    public void before() {
+        mockTracer.reset();
     }
+
+    @Test
+    public void testTracerWrapped() {
+        assertNotEquals(MockTracer.class, tracer.getClass());
+
+        tracer.buildSpan("testop").startManual();
+
+        Mockito.verify(tracerObserver).onStart(Matchers.any(SpanData.class));
+    }
+
 }
